Split keyboard and mouse drivers and fix boot loader
