+++ kernel_main.c
@@ -1,5 +1,7 @@
 #include <stdint.h>
 #include "hardware.h"
+#include "keyboard.h"
+#include "mouse.h"
 #include "fabric.h"
 #include "vga.h"
 // --- pmm.c prototypes ---
